<cstdint> was not included

+++ Engine/Graphics/Include/Luma/Rendering/BufferUtils.h
@@ -1,4 +1,5 @@
 ﻿#pragma once
+#include <cstdint>
 
 namespace Luma
 {
@@ -7,17 +8,17 @@ namespace Luma
 
     namespace BufferUtils
     {
-        IBuffer* createStagingBuffer(IRenderDevice* device, const void* data, size_t size);
-        IBuffer* createVertexBuffer(IRenderDevice* device, const void* data, size_t size);
-        IBuffer* createIndexBuffer(IRenderDevice* device, const void* data, size_t size);
+        IBuffer* createStagingBuffer(IRenderDevice* device, const void* data, uint64_t size);
+        IBuffer* createVertexBuffer(IRenderDevice* device, const void* data, uint64_t size);
+        IBuffer* createIndexBuffer(IRenderDevice* device, const void* data, uint64_t size);
 
-        template<typename T, size_t N>
+        template<typename T, uint64_t N>
         IBuffer* createVertexBuffer(const IRenderDevice* device, const T(&data)[N])
         {
             return createVertexBuffer(device, data, sizeof(data[0]) * N);
         }
 
-        template<typename T, size_t N>
+        template<typename T, uint64_t N>
         IBuffer* createIndexBuffer(const IRenderDevice* device, const T(&data)[N])
         {
             return createIndexBuffer(device, data, sizeof(data[0]) * N);
